Add category policy and auth checks

+++ src/Controller/CategoriesController.php
@@ -53,6 +53,17 @@ class CategoriesController extends AppController
     public function add()
     {
         $category = $this->Categories->newEmptyEntity();
+
+        // Get the identity from the request
+        $user = $this->request->getAttribute('identity');
+
+        $authCheck = $user->can('delete', $category);
+
+        if (!$authCheck) {
+            $this->Flash->error(__('Only admin can add categories'));
+            return $this->redirect(['action' => 'index']);
+        }
+
         if ($this->request->is('post')) {
             $category = $this->Categories->patchEntity($category, $this->request->getData());
             if ($this->Categories->save($category)) {
@@ -77,6 +88,16 @@ class CategoriesController extends AppController
     {
         $category = $this->Categories->get($id, contain: ['BoardGames']);
 
+        // Get the identity from the request
+        $user = $this->request->getAttribute('identity');
+
+        $authCheck = $user->can('edit', $category);
+
+        if (!$authCheck) {
+            $this->Flash->error(__('Only admin can edit categories'));
+            return $this->redirect(['action' => 'view', $category->id]);
+        }
+
         if ($this->request->is(['patch', 'post', 'put'])) {
             $category = $this->Categories->patchEntity($category, $this->request->getData());
             if ($this->Categories->save($category)) {
@@ -100,6 +121,19 @@ class CategoriesController extends AppController
     public function delete($id = null)
     {
         $this->request->allowMethod(['post', 'delete']);
+
+        $category = $this->Categories->get($id, contain: ['BoardGames']);
+
+        // Get the identity from the request
+        $user = $this->request->getAttribute('identity');
+
+        $authCheck = $user->can('delete', $category);
+
+        if (!$authCheck) {
+            $this->Flash->error(__('Only admin can delete categories'));
+            return $this->redirect(['action' => 'view', $category->id]);
+        }
+
         $category = $this->Categories->get($id);
         if ($this->Categories->delete($category)) {
             $this->Flash->success(__('The category has been deleted.'));
